refactor: split inbox message cell
